+++ magda-web-client/src/Components/Header/HeaderMobile.js
@@ -3,6 +3,8 @@ import { Link } from "react-router-dom";
 import HeaderNav from "./HeaderNav";
 import { config } from "../../config.js";
 import mobileMenu from "../../assets/mobile-menu.svg";
+import mobileMenuClosed from "../../assets/mobile-menu-close.svg";
+
 import govtLogo from "../../assets/au-govt-logo-mobile.svg";
 
 class HeaderMobile extends Component {
@@ -36,7 +38,14 @@ class HeaderMobile extends Component {
                         className="mobile-toggle"
                         onClick={() => this.toggleMenu()}
                     >
-                        <img src={mobileMenu} alt="open menu" />
+                        <img
+                            src={`${
+                                this.state.isMobileMenuOpen
+                                    ? mobileMenuClosed
+                                    : mobileMenu
+                            }`}
+                            alt="open menu"
+                        />
                     </button>
                 </div>
                 <div